finish commit

+++ src/components/Posts/DeletePost.jsx
@@ -6,8 +6,12 @@ import { deletePost } from "../../redux/reducers/posts";
 function DeletePost({ post }) {
   const dispatch = useDispatch();
 
+  const profile = useSelector((state) => state.authorization.profile);
+
   const handleDelete = () => {
-    dispatch(deletePost(post.id));
+    if (profile.id === post.userId) {
+      dispatch(deletePost(post.id));
+    }
   };
 
   const authorizing = useSelector((state) => state.authorization.authorizing);
